refactor: hoist cmd_sessions out of main() into sessions_cmd.py
